# sql: add pprof labels based on the statement type