pprof profiles will now be enhanced with the type of the statement that
caused the creation of each stack. statement types are things like
'SELECT', 'INSERT', 'ALTER TABLE', etc.

This example output is produced based on a kv95 workload:

```
(pprof) tags
 stmt: Total 27.7s
       25.4s (91.77%): SELECT
        2.3s ( 8.23%): INSERT
```

Release note: None