@jordanlewis 2018-10-03
Here's a sample profile created with the new tags:

[pprof.samples.cpu.007.pb.gz](https://github.com/cockroachdb/cockroach/files/2443336/pprof.samples.cpu.007.pb.gz)