@jordanlewis 2018-10-03
https://rakyll.org/profiler-labels/ has more details on how the labels work,  btw.