@tschottdorf 2018-10-03
Just a drive by question, what's the cost of a label? Are they cheap enough to throw in basically as we please?