@jordanlewis 2018-10-03
They seem very cheap, but I don't know about using them more than on the order of once per query. They do a `context.Value` call and call a runtime function that sets a field on a pointer.