@jordanlewis 2018-10-03
```
Switching to pprof-labels~
+ make bench PKG=./pkg/bench BENCHTIMEOUT=5m 'BENCHES=BenchmarkSelect1/^Cockroach$$' 'TESTFLAGS=-count 10 -benchmem'
Switching to pprof-labels
+ make bench PKG=./pkg/bench BENCHTIMEOUT=5m 'BENCHES=BenchmarkSelect1/^Cockroach$$' 'TESTFLAGS=-count 10 -benchmem'
name                 old time/op    new time/op    delta
Select1/Cockroach-8     126µs ± 1%     140µs ±13%  +10.74%  (p=0.000 n=9+9)

name                 old alloc/op   new alloc/op   delta
Select1/Cockroach-8    18.8kB ± 0%    19.2kB ± 0%   +2.22%  (p=0.000 n=7+9)

name                 old allocs/op  new allocs/op  delta
Select1/Cockroach-8       164 ± 0%       169 ± 0%   +3.05%  (p=0.000 n=10+10)
```

Pretty disappointing. Not sure this is worth it. Though the `Do` call is high level and always allocates something - let me see if I can improve the performance by doing things with lower level calls.