@jordanlewis 2018-10-03
Replacing the `Do` call makes this seem ok again.

```
name                 old time/op    new time/op    delta
Select1/Cockroach-8     126µs ± 2%     127µs ± 0%    ~     (p=0.156 n=10+9)

name                 old alloc/op   new alloc/op   delta
Select1/Cockroach-8    18.8kB ± 0%    19.2kB ± 0%  +2.26%  (p=0.000 n=10+10)

name                 old allocs/op  new allocs/op  delta
Select1/Cockroach-8       164 ± 0%       169 ± 0%  +3.05%  (p=0.000 n=10+10)
```