@nvanbenschoten 2018-10-03
Can we statically allocate the labels for all statement types to avoid the increase in memory usage?