@jordanlewis 2018-10-03
We'll still have to allocate a map in `WithLabels`, since the `labelContextKey` isn't exported for us to muck with, but I think you're right that we can avoid making `LabelSet`s over and over. Good suggestion.