@nvanbenschoten 2018-10-04
> Nevermind, there's no way we can reduce allocations here.

Can't we at least avoid the `pprof.Labels` allocation?