@nvanbenschoten 2018-10-04
Really? Aren't there varargs and a slice in `pprof.Labels`?