@jordanlewis 2018-10-05
I ran this change through pprof and didn't see any allocations on those lines. Maybe I did something wrong?