@tbg 2019-02-06
Per our discussion yesterday, I think that if you gated this behind a cluster setting and used the anonymized statement string (on top of a tag label) this would be a very useful addition that we could use in the 2.2 cycle to find out whether we'll want to make larger investments here like the ones you're dreaming of (like adding general byte counter profiles etc that help us profile I/O). WDYT?

PS: for the cluster setting, perhaps the right pattern is to make the setting a deadline at which the labels deactivate -- intended usage would be something like `SET CLUSTER SETTING server.pprof.statements = NOW() + INTERVAL '10 minutes';`.